Tweaks for OS X

+++ src/gpu.cc
@@ -160,7 +160,7 @@ CGLContextObj cgl_init()
     };
     CGLPixelFormatObj pixel_format = NULL;
     CGLContextObj cgl_context = NULL;
-    long int pixel_formats = 0L;
+    GLint pixel_formats = 0L;
     CGLError cgl_error;
